% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_intervals}
\alias{error_intervals}
\title{Error intervals}
\usage{
error_intervals(
  value,
  standard_error,
  margin = 1.96,
  line_width = 0.5,
  color = "black"
)
}
\arguments{
\item{value}{(\code{character}) Name of column to take the reference
value for the error intervals.}

\item{standard_error}{(\code{character}) Name of column to take the
standard error for the error intervals.}

\item{margin}{(\code{numeric(1)}) Margin to use for the error intervals. Margin
applies on both sides. Error bar goes from
\code{value - margin * standard_error} to \code{value + margin * standard_error}.
\code{1.96} by default.}

\item{line_width}{(\code{numeric(1)}) Width of the error bar. \code{0.5} by default.}

\item{color}{(\code{character(1)}) Color to use for the error bar. \code{"black"} by
default.}
}
\value{
A \code{ggplot2::geom_errorbar} object.
}
\description{
Add error intervals to a plot.
}
\examples{
set.seed(1)

data <- data.frame(
  Letter = LETTERS[seq(5)],
  Value = rnorm(5, 20, 2),
  Value_SE = rnorm(5)
)

bar_plot(data, "Letter", "Value", fill_by = "Letter") +
  error_intervals("Value", "Value_SE")
}
