% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_plot.R
\name{hex_plot}
\alias{hex_plot}
\title{Hexagonal binning plot}
\usage{
hex_plot(
  data,
  x,
  y,
  bins = 30,
  facet_row = NULL,
  facet_col = NULL,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  low_color = "#132B43",
  high_color = "#56B1F7",
  theme = "light",
  y_breaks_num = 10,
  x_breaks_num = 10,
  font_size = 15,
  x_angle = 0,
  alpha = 1,
  with_legend = TRUE,
  horizontal = FALSE
)
}
\arguments{
\item{data}{(\code{data.frame}) Data to use for plot.}

\item{x}{(\code{quote} | \code{character}) Name of column to use for x-axis.}

\item{y}{(\code{quote} | \code{character}) Name of column to use for y-axis.}

\item{bins}{(\code{numeric(1)}) Number of bins to use in both vertical and
horizontal directions. \code{30} by default.}

\item{facet_row}{(\code{quote} | \code{character}) Name of column to use for row
faceting. \code{NULL} by default.}

\item{facet_col}{(\code{quote} | \code{character}) Name of column to use for column
faceting. \code{NULL} by default.}

\item{title}{(\code{character(1)}) Title of the plot. \code{NULL} by default.}

\item{x_label}{(\code{character(1)}) Label for x-axis. \code{NULL} by default which
uses the name provided in \code{x}.}

\item{y_label}{(\code{character(1)}) Label for y-axis. \code{NULL} by default which
uses the name provided in \code{y}.}

\item{low_color}{(\code{character(1)}) Gradient color to use for low values.
\verb{#132B43} (blue) by default.}

\item{high_color}{(\code{character(1)}) Gradient color to use for high values.
\verb{#56B1F7} (light blue) by default.}

\item{theme}{(\code{character(1)}) Theme to use for the plot. The options are:
\code{"light"},  \code{"dark_grey"}, \code{"dark_blue"} and \code{"transparent"}. \code{light} by
default.}

\item{y_breaks_num}{(\code{numeric(1)}) Number of breaks to use for y-axis. \code{10}
by default.}

\item{x_breaks_num}{(\code{numeric(1)}) Number of breaks to use for x-axis. \code{10}
by default.}

\item{font_size}{(\code{numeric(1)}) Font size to use for the plot. \code{15} by
default.}

\item{x_angle}{(\code{numeric(1)}) Angle to use for x-axis labels. \code{0}
(horizontal) by default.}

\item{alpha}{(\code{numeric(1)}) Alpha value (transparency) to use for elements in
the plot. \code{0.9} by default.}

\item{with_legend}{(\code{logical(1)}) Whether to include a legend in the plot.
Only used when \code{fill_by} is not \code{NULL}. \code{TRUE} by default.}

\item{horizontal}{(\code{logical(1)}) Whether to plot the elements horizontally.
\code{FALSE} by default.}
}
\value{
A (\code{ggplot2}) plot object.
}
\description{
Elegant and easy to implement hexagonal binning plots.
}
\examples{
# Like a scatter plot, but with hexagons instead of points to represent the
# density of points in a given area.
hex_plot(iris, x = Sepal.Length, y = Petal.Length)

# With a facet for each species
hex_plot(iris, x = Sepal.Length, y = Petal.Length, facet_row = Species)

# With custom gradient colors
hex_plot(
  iris,
  x = Sepal.Length,
  y = Petal.Length,
  low_color = "white",
  high_color = "red"
)
}
\seealso{
Other plots: 
\code{\link{bar_plot}()},
\code{\link{box_plot}()},
\code{\link{hist_plot}()},
\code{\link{line_plot}()},
\code{\link{scatter_plot}()}
}
\concept{plots}
