% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_plot.R
\name{hist_plot}
\alias{hist_plot}
\title{Histogram plot}
\usage{
hist_plot(
  data,
  x,
  bins = 20,
  fill_by = NULL,
  facet_row = NULL,
  facet_col = NULL,
  facet_wrap = NULL,
  title = NULL,
  x_label = NULL,
  y_label = "count",
  color = "#386cb0",
  fill_colors = BREWER_COLORS,
  theme = "light",
  x_breaks_num = 10,
  y_breaks_num = 10,
  font_size = 15,
  x_angle = 0,
  alpha = 0.7,
  with_legend = TRUE,
  horizontal = FALSE
)
}
\arguments{
\item{data}{(\code{data.frame}) Data to use for plot.}

\item{x}{(\code{character}) Name of column to use for x-axis.}

\item{bins}{(\code{numeric(1)}) Number of bins to use. \code{20} by default.}

\item{fill_by}{(\code{character}) Name of column to use to fill the
elements of the plot. If \code{NULL}, the plot will be filled with a single
color. \code{NULL} by default.`}

\item{facet_row}{(\code{character}) Name of column to use for row
faceting. \code{NULL} by default.}

\item{facet_col}{(\code{character}) Name of column to use for column
faceting. \code{NULL} by default.}

\item{facet_wrap}{(\code{character}) Name of column to use for automatic
faceting. \code{NULL} by default.}

\item{title}{(\code{character(1)}) Title of the plot. \code{NULL} by default.}

\item{x_label}{(\code{character(1)}) Label for x-axis. \code{NULL} by default which
uses the name provided in \code{x}.}

\item{y_label}{(\code{character(1)}) Label for y-axis. \code{"count"} by default.}

\item{color}{(\code{character(1)}) Color to use for the plot. Only used when
\code{fill_by} is \code{NULL}. \verb{#386cb0} by default.}

\item{fill_colors}{(\code{character}) Colors to use for filling the plot. Only
used if \code{fill_by} is not \code{NULL}. \code{BREWER_COLORS} by default.}

\item{theme}{(\code{character(1)}) Theme to use for the plot. The options are:
\code{"light"},  \code{"dark_grey"}, \code{"dark_blue"} and \code{"transparent"}. \code{light} by
default.}

\item{x_breaks_num}{(\code{numeric(1)}) Number of breaks to use for x-axis. \code{10}
by default.}

\item{y_breaks_num}{(\code{numeric(1)}) Number of breaks to use for y-axis. \code{10}
by default.}

\item{font_size}{(\code{numeric(1)}) Font size to use for the plot. \code{15} by
default.}

\item{x_angle}{(\code{numeric(1)}) Angle to use for x-axis labels. \code{0}
(horizontal) by default.}

\item{alpha}{(\code{numeric(1)}) Alpha value (transparency) to use for elements in
the plot. \code{0.7} by default.}

\item{with_legend}{(\code{logical(1)}) Whether to include a legend in the plot.
Only used when \code{fill_by} is not \code{NULL}. \code{TRUE} by default.}

\item{horizontal}{(\code{logical(1)}) Whether to plot the elements horizontally.
\code{FALSE} by default.}
}
\value{
A (\code{ggplot2}) plot object.
}
\description{
Elegant and easy to implement histogram plots.
}
\examples{
# For a single variable
hist_plot(iris, x = "Sepal.Length")

# Desagregated by a categorical variable
hist_plot(iris, x = "Sepal.Width", fill_by = "Species")

# With a facet for each species
hist_plot(iris, x = "Petal.Length", fill_by = "Species", facet_row = "Species")
}
\seealso{
Other plots: 
\code{\link{bar_plot}()},
\code{\link{box_plot}()},
\code{\link{hex_plot}()},
\code{\link{line_plot}()},
\code{\link{scatter_plot}()}
}
\concept{plots}
