% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Line plot}
\usage{
line_plot(
  data,
  x,
  y,
  fill_by = NULL,
  facet_row = NULL,
  facet_col = NULL,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  with_points = TRUE,
  color = "#386cb0",
  fill_colors = BREWER_COLORS,
  line_width = 2,
  theme = "light",
  y_breaks_num = 10,
  x_breaks_num = 10,
  font_size = 15,
  x_angle = 0,
  alpha = 1,
  with_legend = TRUE,
  horizontal = FALSE
)
}
\arguments{
\item{data}{(\code{data.frame}) Data to use for plot.}

\item{x}{(\code{quote} | \code{character}) Name of column to use for x-axis.}

\item{y}{(\code{quote} | \code{character}) Name of column to use for y-axis.}

\item{fill_by}{(\code{quote} | \code{character}) Name of column to use to fill the
elements of the plot. If \code{NULL}, the plot will be filled with a single
color. \code{NULL} by default.`}

\item{facet_row}{(\code{quote} | \code{character}) Name of column to use for row
faceting. \code{NULL} by default.}

\item{facet_col}{(\code{quote} | \code{character}) Name of column to use for column
faceting. \code{NULL} by default.}

\item{title}{(\code{character(1)}) Title of the plot. \code{NULL} by default.}

\item{x_label}{(\code{character(1)}) Label for x-axis. \code{NULL} by default which
uses the name provided in \code{x}.}

\item{y_label}{(\code{character(1)}) Label for y-axis. \code{NULL} by default which
uses the name provided in \code{y}.}

\item{with_points}{(\code{logical(1)}) Whether to add points to the plot. \code{TRUE}
by default.}

\item{color}{(\code{character(1)}) Color to use for the plot. Only used when
\code{fill_by} is \code{NULL}. \verb{#386cb0} by default.}

\item{fill_colors}{(\code{character}) Colors to use for filling the plot. Only
used if \code{fill_by} is not \code{NULL}. \code{BREWER_COLORS} by default.}

\item{line_width}{(\code{numeric(1)}) Width of the line. \code{2} by default.}

\item{theme}{(\code{character(1)}) Theme to use for the plot. The options are:
\code{"light"},  \code{"dark_grey"}, \code{"dark_blue"} and \code{"transparent"}. \code{light} by
default.}

\item{y_breaks_num}{(\code{numeric(1)}) Number of breaks to use for y-axis. \code{10}
by default.}

\item{x_breaks_num}{(\code{numeric(1)}) Number of breaks to use for x-axis. \code{10}
by default.}

\item{font_size}{(\code{numeric(1)}) Font size to use for the plot. \code{15} by
default.}

\item{x_angle}{(\code{numeric(1)}) Angle to use for x-axis labels. \code{0}
(horizontal) by default.}

\item{alpha}{(\code{numeric(1)}) Alpha value (transparency) to use for elements in
the plot. \code{0.9} by default.}

\item{with_legend}{(\code{logical(1)}) Whether to include a legend in the plot.
Only used when \code{fill_by} is not \code{NULL}. \code{TRUE} by default.}

\item{horizontal}{(\code{logical(1)}) Whether to plot the elements horizontally.
\code{FALSE} by default.}
}
\value{
A (\code{ggplot2}) plot object.
}
\description{
Elegant and easy to implement line plots.
}
\examples{
# Normal line plot
line_plot(iris, x = Petal.Width, y = Sepal.Length)

# With different colors for each species
line_plot(iris, x = Petal.Width, y = Sepal.Length, fill_by = Species)

# With different row facets for each species
line_plot(iris, x = Petal.Width, y = Sepal.Length, facet_row = Species)
}
\seealso{
Other plots: 
\code{\link{bar_plot}()},
\code{\link{box_plot}()},
\code{\link{hex_plot}()},
\code{\link{hist_plot}()},
\code{\link{scatter_plot}()}
}
\concept{plots}
